refactored models

+++ src/models/PaymentMethodModel.cs
@@ -15,7 +15,7 @@ public class PaymentMethodModel
 
     [Required(ErrorMessage = "Card number is required.")]
     [Range(100000000000, 9999999999999999, ErrorMessage = "Card number must be between 12 to 16 digits")]
-    public required decimal CardNumber { get; set; }
+    public required double CardNumber { get; set; }
 
     [Required(ErrorMessage = "Cardholder name is required.")]
     [RegularExpression(@"^[a-zA-Z\s]+$", ErrorMessage = "Cardholder name can only contain letters and spaces.")]
